aiohttp/requirements.txt
aiodns==2.0.0
aiohttp==3.5.4
aiohttp-jinja2==1.1.2
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
cchardet==2.1.4
cffi==1.12.3
chardet==3.0.4
elasticsearch==7.0.2
elasticsearch-async==6.2.0
idna==2.8
idna-ssl==1.1.0
importlib-metadata==0.18
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.2.0
multidict==4.5.2
packaging==19.0
pluggy==0.12.0
py==1.8.0
pycares==3.0.0
pycparser==2.19
pyparsing==2.4.0
pytest==5.0.1
pytest-aiohttp==0.3.0
PyYAML==5.1.1
six==1.12.0
typing==3.7.4
typing-extensions==3.7.4
ujson==1.35
urllib3==1.25.3
uvloop==0.12.2
wcwidth==0.1.7
yarl==1.3.0
zipp==0.5.2
